moved game loading to Game class
