+++ engine/src/gameloop.ts
@@ -1,5 +1,8 @@
-import { Renderer } from "./renderer/renderer.js";
+import { Renderer, SpriteLayer } from "./renderer/renderer.js";
 import { Updatable } from "./types.js";
+import { SceneManager } from "./sceneManager.js";
+import { ImportManager } from "./importManager.js";
+import { PostProcessing } from "./engineExport.js";
 
 const nwjs = require("nw.gui");
 
@@ -22,7 +25,7 @@ export class Game {
         this.updateQueue.push(object);
     }
 
-    static gameloop(t: number) {
+    private static gameloop(t: number) {
         const now = t * 0.001;
         const dt = Math.min(0.1, now - this.time);
         this.time = now;
@@ -33,6 +36,47 @@ export class Game {
 
         if (this.running) {
             requestAnimationFrame(this.gameloop.bind(this));
+        }
+    }
+
+    /**
+     * Load a project from a project.json file
+     * 
+     * @param src the path to the folder containing `project.json`
+     */
+    static loadGame(src: string = "./") {
+        const fs = require("fs");
+
+        // load the project.json which contains all info needed to initialise the game
+        const data = fs.readFileSync(src + 'project.json', {encoding: "utf-8"});
+        const game_data = JSON.parse(data.toString());
+    
+        Renderer.setResolution(game_data["resolution"][0], game_data["resolution"][1]);
+        this.loadLayers(game_data["layers"]);
+        this.loadTextures(game_data["textures"]);
+        this.loadShaders(game_data["shaders"]);
+    
+        Game.addToUpdateQueue(SceneManager);
+        SceneManager.loadScene(game_data["start_scene"]);
+    }
+    
+    private static loadLayers(layers: string[]) {
+        for (const layer of layers) {
+            Renderer.addLayer(new SpriteLayer(), layer);
+        }
+    }
+    
+    private static loadTextures(textures: {name: string, path: string}[]) {
+        for (const texture of textures) {
+            Renderer.loadTexture(texture["path"], texture["name"]);
+        }
+    }
+    
+    private static loadShaders(shaders: {name: string, args: any[]}[]) {
+        for (const shader of shaders) {
+            const typed_constr = ImportManager.getShader(shader["name"]);
+            const args = typed_constr.parseArgs(shader["args"]);
+            PostProcessing.add(new typed_constr.constr(...args));
         }
     }
